requirements.txt
pyyaml
pandas
openai
requests
tqdm